Add in the first Listen function
This is the start of modularizing the code a little bit better
so that I can add in the code to control the beacon/etc as well.

This is done in this CL by building a function that'll block
for a set period of time and listen for an incoming message.  This
function should be very useful to easily listen for messages
in between blinking out morse code.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>