Getting the Beacon back up and running with the new library
This is a little tricky.  To turn the Beacon to a differnt color
when a handhake/data is detected I use a callback.  I'm not sure
if this is the best solution, but it seems to work great, and it
allows me to completely decouple the beacon control code from the
laser-message-receiving code.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>