First commit w/ Morse code & Laser Receiving working!
This is the first time they're both working together.  The morse code
blinking program uses a "Listen while waiting" function to generate
the dots and dashes for the morse code.  Thus, it's listening for an
incoming message the vast majority of the time it's transmitting.  It
seems to work great!

Next up -- using those incoming messages for something other than just
printing to the terminal.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>