Better memory management for LaserReceiver
Previously a new chunk of memory was being malloc'd for the incoming
message.  It's probably safer to have a preallocated buffer so we can
see just how much memory we're using. This also allows us to use the
same buffer for the Receiver object as well as the client, which
is important for allowing the longest messages possible.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>