Store the LaserReceiver object on the stack
This required a couple of minor changes to make possible, but basically
I wanted to put that object on the stack instead of the heap with "new"
like I was doing before.  For one thing, I know I need extactly 1 of them
forever, so it seems like the obvious choice, but also it allows me
to get a better idea of how much memory I actually have to play with since
it's all defined at compile time now.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>