Get store/restore morse messages from EEPROM
In this version messages that come in on the laser are stored into EEPROM
and then used as the message to display on Morse code.  It's finally working
end-to-end.  Phew.  Not done, but with this code you can actually update the
morse code beacon from a distance with a new message!

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>